Establish report class dictionary only once
